+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+- There is now a `GibsonLanniPSF` for modeling realistic 3D PSF's and
+  that can account for details such as sample, coverslip, and
+  immersion media refractive indexes and thicknesses.
+- A `PSFBuilder` interface was added to allow for a PSF to be built at
+  various times during the simulation, rather than all at once. This
+  addition is necessary to account for axial stage positions that
+  might change during the simulation.
+    
 ## [v0.4.0]
 Contains [ALICA v0.2.1]